requirements.txt
astroid==3.2.4
certifi==2025.8.3
charset-normalizer==3.4.3
dill==0.4.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
pylint==3.2.7
python-dotenv==1.0.1
requests==2.32.4
ruff==0.12.9
tomli==2.2.1
tomlkit==0.13.3
typing-extensions==4.13.2
urllib3==2.2.3
